% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfname.R
\name{name_in}
\alias{name_in}
\alias{name_out}
\title{Name and unname data frames}
\usage{
name_in(d, col_name = "name")

name_out(d, col_name)
}
\arguments{
\item{d}{A data.frame}

\item{col_name}{For \code{name_in}, the name of the column to which rownames
will be copied. For \code{name_out}, the name of the column to move into
rownames.}
}
\description{
Return a data frame with rownames moved into a named column, or vice versa.
}
\examples{
unnamed_cars <- name_in(mtcars, "car_name")

cars <- name_out(unnamed_cars, "car_name")
}

